Added special handling of padding only cases for the resample tool. Added handling of out of memory exceptions to the crop tool. Fixed a compiler warning.

+++ src/QtUtils/Bridge/detail/QtComboBoxConnector.cc
@@ -218,7 +218,7 @@ void QtComboBoxConnector::UpdateComboBoxItems( QPointer< QtComboBoxConnector > q
     else
     {
       combobox->clear();
-      for ( size_t i = 0; i < num_items; i++ )
+      for ( int i = 0; i < num_items; i++ )
       {
         combobox->addItem( QString( option_list[ i ].c_str() ) );
       }
